Some database testing
When starting, the App now requests a list of all available objects from a Postgrest server.
When tapping with two fingers at the same time it downloads the first object from the list from a webserver and from then on places this object (a lantern) instead of the standard cube when tapping with one finger.